moved error data to separate file, removed CCfault
