+++ MSXIV_Driver_Display/lib/widgets/errors.dart
@@ -1,66 +1,6 @@
 import 'package:MSXIV_Driver_Display/constants/std_colors.dart';
+import 'package:MSXIV_Driver_Display/utils/errors.dart';
 import 'package:flutter/material.dart';
-
-const int BPS_STATE_FAULT_KILLSWITCH = (1 << 0);
-const int BPS_STATE_FAULT_AFE_CELL = (1 << 1);
-const int BPS_FAULT_SOURCE_AFE_TEMP = (1 << 2);
-const int BPS_STATE_FAULT_AFE_FSM = (1 << 3);
-const int BPS_STATE_FAULT_RELAY = (1 << 4);
-const int BPS_STATE_FAULT_CURRENT_SENSE = (1 << 5);
-const int BPS_STATE_FAULT_ACK_TIMEOUT = (1 << 6);
-
-enum EEChargerFault {
-  EE_CHARGER_FAULT_HARDWARE_FAILURE,
-  EE_CHARGER_FAULT_OVER_TEMP,
-  EE_CHARGER_FAULT_WRONG_VOLTAGE,
-  EE_CHARGER_FAULT_POLARITY_FAILURE,
-  EE_CHARGER_FAULT_COMMUNICATION_TIMEOUT,
-  EE_CHARGER_FAULT_CHARGER_OFF,
-  NUM_EE_CHARGER_FAULTS,
-}
-
-enum EESolarFault {
-  EE_SOLAR_FAULT_MCP3427,
-  EE_SOLAR_FAULT_MPPT_OVERCURRENT,
-  EE_SOLAR_FAULT_MPPT_OVERVOLTAGE,
-  EE_SOLAR_FAULT_MPPT_OVERTEMPERATURE,
-  EE_SOLAR_FAULT_OVERCURRENT,
-  EE_SOLAR_FAULT_NEGATIVE_CURRENT,
-  EE_SOLAR_FAULT_OVERVOLTAGE,
-  EE_SOLAR_FAULT_OVERTEMPERATURE,
-  NUM_EE_SOLAR_FAULTS,
-}
-
-enum ErrorStates {
-  CentreConsoleFault,
-  MCIAckFailed,
-  PedalACKFail,
-  CentreConsoleStateTransitionFault,
-  ChargerFaultHardwareFailure,
-  ChargerFaultOverTemperature,
-  ChargerFaultWrongVoltage,
-  ChargerFaultPolarityFailure,
-  ChargerFaultCommunicationTimeout,
-  ChargerFaultChargerOff,
-  SolarFaultMCP3427,
-  SolarFaultMPPTOverCurrent,
-  SolarFaultMPPTOverVoltage,
-  SolarFaultMPPTOverTemperature,
-  SolarFaultOverCurrent,
-  SolarFaultNegativeCurrent,
-  SolarFaultOverVoltage,
-  SolarFaultOverTemperature,
-  BPSACKFailed,
-  BPSKillSwitch,
-  BPSAFECellFault,
-  BPSAFETempFault,
-  BPSAFEFSMFault,
-  BPSRelayFault,
-  BPSCurrentSenseFault,
-  // TODO: Implement BMSOvervoltage and MCIOverTemp
-  BMSOverVoltage,
-  MCIOverTemp,
-}
 
 class Errors extends StatelessWidget {
   final List<ErrorStates> errors;
@@ -75,8 +15,6 @@ class Errors extends StatelessWidget {
       child: Column(
         mainAxisAlignment: MainAxisAlignment.start,
         children: <Widget>[
-          if (errors.contains(ErrorStates.CentreConsoleFault))
-            ErrorItem("Centre Console", "Fault"),
           if (errors.contains(ErrorStates.MCIAckFailed))
             ErrorItem("MCI", "ACK Failed"),
           if (errors.contains(ErrorStates.PedalACKFail))
